requirements.txt
python-dotenv
opencv-python
shapely
openai
matplotlib
requests
multiprocess
numpy<2
scipy
joblib